refactor: extract Threads to :common:android
Part of moving theming to :common:android, so we can make
CollectionManager common-compatible

Part of 20737
