+++ common/android/src/main/java/com/ichi2/anki/common/utils/android/Threads.kt
@@ -14,7 +14,7 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.ichi2.utils
+package com.ichi2.anki.common.utils.android
 
 import android.os.Looper
 import java.lang.RuntimeException
